requirements.txt
arrow==0.17.0
asgiref==3.4.1
b2==3.0.3
b2sdk==1.12.0
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.7
django-cors-headers==3.9.0
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
docutils==0.16
funcsigs==1.0.2
gunicorn==20.1.0
idna==3.2
logfury==0.1.2
phx-class-registry==3.0.5
Pillow==8.3.2
postmarker==0.18.2
psycopg2==2.9.1
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
rst2ansi==0.1.5
six==1.16.0
sqlparse==0.4.2
tqdm==4.62.3
urllib3==1.26.7
wheel==0.37.0
whitenoise==5.3.0
